last day versionn 1

+++ htproject/src/main/java/com/project/htproject/response/LoginMessage.java
@@ -4,7 +4,7 @@ public class LoginMessage {
 
 	private String message;
 	private boolean status;
-	private Long Id;
+	private int Id;
 	public String getMessage() {
 		return message;
 	}
@@ -17,13 +17,13 @@ public class LoginMessage {
 	public void setStatus(boolean status) {
 		this.status = status;
 	}
-	public Long getId() {
+	public int getId() {
 		return Id;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		Id = id;
 	}
-	public LoginMessage(String message, boolean status, Long id) {
+	public LoginMessage(String message, boolean status, int id) {
 		super();
 		this.message = message;
 		this.status = status;
@@ -33,6 +33,7 @@ public class LoginMessage {
 	public String toString() {
 		return "LoginMessage [message=" + message + ", status=" + status + ", Id=" + Id + "]";
 	}
+	
 	
 	
 }
